refactor: move eval harness under cmd
Move the evaluation framework and suite packages from the repository-level internal tree into cmd/eval/internal/eval so benchmark harness code is scoped to the eval command family.

Update helper imports and schema fixture paths for the new package depth, and add a source-layout guard that prevents the eval harness from returning to the shared internal boundary.
