+++ cmd/eval/internal/eval/readbench/schema_test.go
@@ -5,6 +5,7 @@ import (
 	"os"
 	"path/filepath"
 	"reflect"
+	"runtime"
 	"strings"
 	"testing"
 )
@@ -57,7 +58,7 @@ func TestResultSchemaIsCheckedInAndMatchesResultFields(t *testing.T) {
 func readResultSchema(t *testing.T) jsonSchema {
 	t.Helper()
 
-	data, err := os.ReadFile(filepath.Join("..", "..", "..", filepath.FromSlash(ResultSchemaPath)))
+	data, err := os.ReadFile(filepath.Join(schemaDir(t), "readbench-result.schema.json"))
 	if err != nil {
 		t.Fatalf("read schema %s: %v", ResultSchemaPath, err)
 	}
@@ -66,6 +67,15 @@ func readResultSchema(t *testing.T) jsonSchema {
 		t.Fatalf("parse schema %s: %v", ResultSchemaPath, err)
 	}
 	return schema
+}
+
+func schemaDir(t *testing.T) string {
+	t.Helper()
+	_, currentFile, _, ok := runtime.Caller(0)
+	if !ok {
+		t.Fatal("runtime.Caller failed")
+	}
+	return filepath.Join(filepath.Dir(currentFile), "..", "..", "..", "schemas")
 }
 
 func resultJSONFields(t *testing.T) (map[string]struct{}, map[string]struct{}) {
