+++ cmd/eval/internal/eval/suites/casmodel/suite.go
@@ -7,11 +7,11 @@ import (
 	"fmt"
 	"time"
 
+	"github.com/dewebprotocol/malt/cmd/eval/internal/eval/framework"
+	"github.com/dewebprotocol/malt/cmd/eval/internal/eval/suites/configjson"
 	"github.com/dewebprotocol/malt/core/cas"
 	casmock "github.com/dewebprotocol/malt/core/cas/mock"
 	"github.com/dewebprotocol/malt/core/metrics"
-	"github.com/dewebprotocol/malt/internal/eval/framework"
-	"github.com/dewebprotocol/malt/internal/eval/suites/configjson"
 	cid "github.com/ipfs/go-cid"
 )
