+++ cmd/eval/internal/eval/suites/proofoverhead/suite_test.go
@@ -8,7 +8,7 @@ import (
 	"path/filepath"
 	"testing"
 
-	"github.com/dewebprotocol/malt/internal/eval/framework"
+	"github.com/dewebprotocol/malt/cmd/eval/internal/eval/framework"
 )
 
 func TestSuiteMeasuresMapAndListProofOverhead(t *testing.T) {
